Server passed to another file
